:refactor: 컨트롤러 인터페이스 분리
